[#5]개인회원 프로필 조회 기능 & 화면 추가 전체 회원 조회 페이지에서 회원 데이터 클릭시 개인 회원 프로필 조회로 이동

+++ src/main/java/com/laundry/laundry/repository/MemberRepository.java
@@ -11,7 +11,7 @@ public interface MemberRepository {
     Member update(Member member);
     Member delete(Long id);
 
-    Optional<Member> findById(Long id);
+    Optional<Member> findById(int id);
     Optional<Member> findByName(String name);
     Optional<Member> findByPhoneNumber(String phoneNumber);
     Optional<Member> findByHomePhoneNumber(String homePhoneNumber);
